renamed AccountCard component to AccountCorgiCard
